separated gui from stuff
