+++ demo/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module com.example.demo {
+    requires javafx.controls;
+    requires javafx.fxml;
+
+    opens com.example.demo to javafx.fxml;
+    exports com.example.demo;
+    requires pl.put.poznan.transformer;
+}
