+++ demo/src/main/resources/com/example/demo/hello-view.fxml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+
+
+<AnchorPane prefHeight="637.0" prefWidth="896.0" xmlns="http://javafx.com/javafx/11.0.1"
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.demo.HelloController">
+    <children>
+        <VBox alignment="CENTER" prefHeight="475.0" prefWidth="896.0" spacing="20.0">
+            <padding>
+                <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
+            </padding>
+            <children>
+                <TableView fx:id="table" prefHeight="332.0" prefWidth="564.0">
+                    <columns>
+                        <TableColumn fx:id="idcol" prefWidth="75.0" text="id"/>
+                        <TableColumn fx:id="namecol" prefWidth="75.0" text="name"/>
+                        <TableColumn fx:id="typecol" prefWidth="75.0" text="type"/>
+                        <TableColumn fx:id="areacol" prefWidth="75.0" text="area"/>
+                        <TableColumn fx:id="cubecol" prefWidth="75.0" text="cube"/>
+                        <TableColumn fx:id="heatingcol" prefWidth="75.0" text="heating"/>
+                        <TableColumn fx:id="lightcol" prefWidth="100.0" text="light"/>
+                        <TableColumn fx:id="light2col" prefWidth="94.0" text="light/m2"/>
+                        <TableColumn fx:id="heating3" prefWidth="95.0" text="heating/m3"/>
+                    </columns>
+                </TableView>
+                <Button fx:id="load" mnemonicParsing="false" onMouseClicked="#onHelloButtonClick" text="Load data"/>
+            </children>
+        </VBox>
+        <TextField fx:id="textField" layoutX="119.0" layoutY="540.0"/>
+        <Label layoutX="10.0" layoutY="545.0" prefHeight="16.0" prefWidth="96.0" text="Heating norm"/>
+        <Button layoutX="311.0" layoutY="540.0" mnemonicParsing="false" onMouseClicked="#heatingCount" text="Apply"/>
+    </children>
+</AnchorPane>
